chore(refactor): Use vulnerability instead of vulnerabiities
