+++ src/Components/PresentationalComponents/CSAwRuleBox/CSAwRuleSummary.js
@@ -6,7 +6,7 @@ import { injectIntl } from 'react-intl';
 import sanitizeHtml from 'sanitize-html';
 import { Truncate } from '@redhat-cloud-services/frontend-components/components/Truncate';
 import { Stack, StackItem, TextContent } from '@patternfly/react-core';
-import { handleCVELink } from '../../../Helpers/VulnerabilitiesHelper';
+import { handleCVELink } from '../../../Helpers/VulnerabilityHelper';
 import messages from '../../../Messages';
 
 /**
