+++ src/Components/SmartComponents/CVEs/CVEs.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Stack, StackItem } from '@patternfly/react-core';
 import { VULNERABILITIES_HEADER, CVES_ALLOWED_PARAMS } from '../../../Helpers/constants';
 import { GenericError } from '../../PresentationalComponents/EmptyStates/EmptyStates';
-import { createCveListByAccount } from '../../../Helpers/VulnerabilitiesHelper';
+import { createCveListByAccount } from '../../../Helpers/VulnerabilityHelper';
 import { constructFilterParameters, updateRef, useCreateUrlParams } from '../../../Helpers/MiscHelper';
 import BusinessRiskModal from '../Modals/BusinessRiskModal';
 import StatusModal from '../Modals/CveStatusModal';
