+++ src/Components/SmartComponents/SystemCves/SystemCves.js
@@ -15,7 +15,7 @@ import { injectIntl, IntlProvider } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 import { SYSTEM_DETAILS_HEADER, CVES_ALLOWED_PARAMS } from '../../../Helpers/constants';
 import { constructFilterParameters } from '../../../Helpers/MiscHelper';
-import { createCveListBySystem } from '../../../Helpers/VulnerabilitiesHelper';
+import { createCveListBySystem } from '../../../Helpers/VulnerabilityHelper';
 import { Stack, StackItem, Text, TextContent, TextVariants } from '@patternfly/react-core';
 import { EmptyVulnerabilityData, GenericError } from '../../PresentationalComponents/EmptyStates/EmptyStates';
 import DownloadReport from '../../../Helpers/DownloadReport';
