+++ src/Helpers/VulnerabilityHelper.js
