+++ src/Helpers/VulnerabilityHelper.test.js
@@ -4,7 +4,7 @@ import {
     createCveListBySystem,
     handleCVELink,
     renderSystemCSAwIcon
- } from './VulnerabilitiesHelper';
+ } from './VulnerabilityHelper';
  
 const { parameters } = initialState;
