+++ src/Helpers/constants.js
@@ -3,7 +3,7 @@ import BellIcon from '@patternfly/react-icons/dist/js/icons/bell-icon';
 import ExclamationCircleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
 import ExclamationTriangleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-triangle-icon';
 import { classNames, expandable, sortable } from '@patternfly/react-table';
-import { renderSystemCSAwIcon } from './VulnerabilitiesHelper';
+import { renderSystemCSAwIcon } from './VulnerabilityHelper';
 import StatusColumn from '../Components/PresentationalComponents/StatusColumn/StatusColumn';
 import messages from '../Messages';
 import { intl } from '../Utilities/IntlProvider';
