+++ src/Store/Selectors/VulnerabilitySelectors.js
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-import { createCveListByAccount } from '../../Helpers/VulnerabilitiesHelper';
+import { createCveListByAccount } from '../../Helpers/VulnerabilityHelper';
 
 const cveListSelector = state => state.VulnerabilitiesStore.cveList;
